requirements.txt
modulegraph~=0.18
macholib~=1.14
py2app~=0.24
setuptools~=56.0.0
altgraph~=0.17
Pillow~=8.2.0
pytesseract~=0.3.7
timecode~=1.3.1
darkdetect~=0.3.0
wxPython~=4.1.1
pandas~=1.2.4